Accumulated Volume wrong type causing unmarshalling errors in the client (#230)
* Accumulated Volume wrong type causing unmarshalling errors in the client

* update test data int -> float for snapshot.min.av

+++ rest/models/snapshot.go
@@ -236,7 +236,7 @@ type LastTradeSnapshot struct {
 
 // MinuteSnapshot is the most recent minute agg for a ticker.
 type MinuteSnapshot struct {
-	AccumulatedVolume     int     `json:"av,omitempty"`
+	AccumulatedVolume     float64 `json:"av,omitempty"`
 	Close                 float64 `json:"c,omitempty"`
 	High                  float64 `json:"h,omitempty"`
 	Low                   float64 `json:"l,omitempty"`
